+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
  - Renamed Microsoft.SurfaceHub project to Microsoft
  - Fixed volume ramping for sharp displays
+ - Only query display state on port assignment if the port is connected
+ - Query displays when port becomes connected
 
 ## [8.0.1] - 2018-09-14
 ### Changed